Separated all components into diffferent file & use prop-types for type-checking
