+++ instagram/src/components/CommentSection/CommentSection.css
@@ -1,0 +1,19 @@
+/*Commment Section*/
+
+.comment-section {
+    height: 100px;
+    border: 3px solid red;
+    overflow-y: scroll;
+    border-bottom: 3px solid #000;
+}
+  
+.comment {
+    height: 20px;
+    border: 3px solid red;
+    font-size: 1.5rem;
+    text-align: left;
+}
+  
+.username {
+    font-weight: bold;
+}
