+++ instagram/src/components/PostContainer/PostContainer.css
@@ -1,0 +1,56 @@
+/*Post Container*/
+.post-wrapper {
+    width: 600px;
+    height: 815px;
+    margin: 0 auto;
+    border: 3px solid #000;
+    margin-bottom: 30px;
+}
+  
+.header-wrapper {
+    display: flex;
+    width: 100%;
+    height: 70px;
+    background-color: #fafafa;
+    border-bottom: 3px solid #000;
+}
+  
+.thumbImg-wrapper {
+    width: 25px;
+    height: 25px;
+    border: 3px solid #000;
+    border-radius: 50%;
+}
+  
+.post-image {
+    width: 100%;
+    height: 500px;
+    border-bottom: 3px solid #000;
+}
+  
+.comment-wrapper {
+    padding: 10px;
+}
+  
+.tot-likes {
+    text-align: left;
+    font-size: 1.5rem;
+    font-weight: bold;
+}
+  
+.newComment-section {
+    display: flex;
+    margin-top: 0;
+    padding: 0;
+    height: 50px;
+}
+  
+.newComment-section > input {
+    display: flex;
+    width: 80%;
+    height: 100%;
+    padding-left: 10px;
+    border: none;
+    outline: 0;
+    font-size: 1.8rem;
+}
