+++ instagram/src/components/SearchBar/SearchBar.css
@@ -1,0 +1,41 @@
+/*Search Bar*/
+header {
+    width: 100%;
+    height: 80px;
+    display: flex;
+    justify-content: space-between; 
+    border-bottom: 2px solid grey;
+    padding: 20px;
+    box-sizing: border-box;
+}
+
+.logo-wrapper {
+    width: 32%;
+    display: flex;
+    /* flex-direction: column; */
+    justify-content: flex-start;
+    align-content: center;
+}
+
+.logo-wrapper > h3 {
+    border-left: 2px solid #000;
+    font-size: 2rem;
+}
+
+.search-wrapper {
+    width: 32%;
+}
+
+.search-bar { 
+    height: 25px;
+    width: 200px;
+    background-color: #fafafa; 
+    border-radius: 5px;
+    text-align: center;
+    font-size: 1.5rem;
+}
+
+.nav-icons {
+    width: 32%;
+    display: flex;
+}
